requirements.txt
aiocontextvars==0.2.2
certifi==2020.12.5
contextvars==2.4
et-xmlfile==1.0.1
immutables==0.15
jdcal==1.4.1
loguru==0.5.3
mysql-connector-python==8.0.23
numpy==1.19.5
openpyxl==3.0.5
pandas==1.1.5
protobuf==3.14.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
PyYAML==5.4.1
schedule==1.0.0
six==1.15.0
SQLAlchemy==1.3.22
